Many changes: - Misc. changes I dont remember. - fixed `TypeOfVar` and `ReturnTypeOf` - jankman, living within the walls of function captures, has been removed...

+++ crates/ospl-vm/src/arena.rs
@@ -16,7 +16,7 @@ pub const MEMMAX: usize = 1024;
 #[derive(Default, Clone)]
 pub struct ArenaItem {
     pub inner: RuntimeValue,
-    pub refcount: usize,
+    pub refcount: u16,  // stupid to make this an usize
     next_free: Option<usize>,
 }
